refactor: extract format as its own file
