+++ src/compile/legend/parse.ts
@@ -23,7 +23,8 @@ import {Legend} from '../../legend';
 import {normalizeTimeUnit} from '../../timeunit';
 import {GEOJSON} from '../../type';
 import {deleteNestedProperty, getFirstDefined, keys, varName} from '../../util';
-import {mergeTitleComponent, numberFormat} from '../common';
+import {mergeTitleComponent} from '../common';
+import {numberFormat} from '../format';
 import {guideEncodeEntry} from '../guide';
 import {isUnitModel, Model} from '../model';
 import {parseGuideResolve} from '../resolve';
